# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roadmap for Jira API",
        "description": "Endpoint mapping for Roadmap for Jira App.",
        "contact": {
            "email": "contact@mrrafael.ca"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:3030",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/v1/user/validate": {
            "put": {
                "tags": [
                    "User"
                ],
                "description": "If found, returns the username associated with the provided username.",
                "operationId": "Validate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserValidationParam"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object with user's name and avatar",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/UserValidationResult"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/sendcode": {
            "put": {
                "tags": [
                    "User"
                ],
                "description": "Generates an authentication code for the specified user.",
                "operationId": "SendCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendCodeParam"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/removeTokenJira": {
            "delete": {
                "tags": [
                    "User"
                ],
                "description": "Remove o token Jira cadastrado para o usuário logado.",
                "operationId": "RemoveTokenJira",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Objeto JSON contendo a mensagem de erro",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorMessageModel"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Objeto JSON contendo a mensagem de erro",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/configTokenJira": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Configures a new Jira token for the logged-in user. v4",
                "operationId": "ConfigTokenJira",
                "requestBody": {
                    "required": true,
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string",
                                "example": "6c71d92e-1af3-4597-a51c-e7689334b7ab"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorMessageModel"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserValidationParam": {
                "required": [
                    "userName"
                ],
                "properties": {
                    "userName": {
                        "type": "string",
                        "example": "john.smith"
                    }
                },
                "type": "object"
            },
            "SendCodeParam": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "john.smith@mybigcompany.com"
                    }
                },
                "type": "object"
            },
            "UserValidationResult": {
                "required": [
                    "name",
                    "avatarUrl",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Smith"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.smith@mybigcompany.com"
                    },
                    "avatarUrl": {
                        "type": "string",
                        "example": "https://r2.1k-cdn.com/sig/size:256/plain/https%3A%2F%2F1k-cdn.com%2Fresimler%2Fyazarlar%2F513202_c27ea_1586708018.jpg"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "required": [
                    "code",
                    "error"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "42S22"
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Complementary Error message"
                    }
                },
                "type": "object"
            },
            "ErrorMessageModel": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}